Add "AWAvenue Ads Rule" to AdGuard Home Filter list
This rule has been added to the "HostlistsRegistry" and is visible in the AdGuard DNS filter list. Hope to add it to the filter list of AdGuard Home to facilitate users to subscribe.

Use Adblock syntax to fight against various advertising SDKs in Android applications from the network level , prevent them from loading.