home: add service ready notification support on Linux
On systemd-based Linux systems, signal service readiness once the web
service is up and running. This allows depending services to accurately
rely on AdGuardHome being active and ready to service.

This feature is particularly useful with Podman, as the container
manager can verify that the service can start correctly after an
auto-update operation.

This new feature does not do anything on non-systemd Linux or other
operating systems.